* include/ruby/ruby.h (rb_event_flag_t):  Maintain integer precision   for clang error (VALUE aka unsigned long vs unsigned int)
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@38084 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/ruby.h
@@ -1584,7 +1584,7 @@ int ruby_native_thread_p(void);
 #define RUBY_EVENT_SPECIFIED_LINE  0x8000
 #define RUBY_EVENT_TRACEPOINT_ALL  0xFFFF
 
-typedef unsigned int rb_event_flag_t;
+typedef unsigned long rb_event_flag_t;
 typedef void (*rb_event_hook_func_t)(rb_event_flag_t evflag, VALUE data, VALUE self, ID mid, VALUE klass);
 
 #define RB_EVENT_HOOKS_HAVE_CALLBACK_DATA 1
